rename youtube scheduler
